LTBENCH-156: Changed timeout properties type to long

+++ test-automation-feign/src/main/kotlin/com/nortal/test/feign/configuration/FeignProperties.kt
@@ -11,26 +11,26 @@ data class FeignProperties(
          * The read timeout is applied to both the TCP socket and for individual read IO operations.
          * The default value is 30 seconds.
          */
-        val readTimeout: Int = 30,
+        val readTimeout: Long = 30,
         /**
          * Sets the default connect timeout for new connections.
          * The connect timeout is applied when connecting a TCP socket to the target host.
          * The default value is 30 seconds.
          */
-        val connectTimeout: Int = 30,
+        val connectTimeout: Long = 30,
         /**
          * Sets the default write timeout for new connections.
          * The write timeout is applied for individual write IO operations.
          * The default value is 30 seconds.
          */
-        val writeTimeout: Int = 30,
+        val writeTimeout: Long = 30,
         /**
          * Sets the default timeout for complete calls.
          * The call timeout spans the entire call: resolving DNS, connecting, writing the request body, server processing, and reading the response body.
          * If the call requires redirects or retries all must complete within one timeout period.
          * The default value is 30 seconds.
          */
-        val callTimeout: Int = 30,
+        val callTimeout: Long = 30,
         /**
          * Configure this client to retry or not when a connectivity problem is encountered.
          * The default value is true.
